made loginUser reusable util
