+++ client/src/pages/Login.js
@@ -1,4 +1,3 @@
-import axios from "axios";
 import React from "react";
 import { useHistory } from "react-router-dom";
 
@@ -19,6 +18,7 @@ import Background from "../assets/login_bg.png";
 import Logo from "../components/Logo";
 import { SET_USER } from "../contexts/types";
 import { UserContext } from "../contexts/UserContext";
+import loginUser from "../utils/loginUser"
 
 function Alert(props) {
   return <MuiAlert elevation={6} variant="filled" {...props} />;
@@ -205,13 +205,3 @@ export default function Login() {
   );
 }
 
-const loginUser = async (user) => {
-  try {
-    const res = await axios.post("/api/auth", user);
-    localStorage.setItem("HatchwayToken", res.data.token);
-    axios.defaults.headers.common["x-auth-token"] = res.data.token;
-    return res.data.user;
-  } catch (err) {
-    console.error(err);
-  }
-};
